Add Publication and VariantPublicationReview
The "Publication" model (and the Publications app) contains information
regarding publications. These are currently assumed to be identifiable
with PubMed IDs, which is used to scrape the abstract from NCBI.

Publications can be associated with multiple variants. For each
such association we may want to store specific information, like
a review summary. So the "VariantPublicationReview" model is this link,
with keys to both Variant and Publication objects.